iris example

+++ DesertLand/ImportDigits.cs
@@ -9,15 +9,15 @@ namespace DesertLand
     public static class ImportDigits
     {
 
-        static float[] Categorical(int[] x)
+        static double[] Categorical(int[] x)
         {
-            var data = x.SelectMany(i => Enumerable.Range(0, 10).Select((v, k) => k == i ? 1f : 0f)).ToArray();
+            var data = x.SelectMany(i => Enumerable.Range(0, 10).Select((v, k) => k == i ? 1.0 : 0.0)).ToArray();
             return data;
         }
 
-        static (float[], float[]) ImportDataset(string[] raw)
+        static (double[], double[]) ImportDataset(string[] raw)
         {
-            var data = raw.Select(l => l.Split(',').Select(float.Parse).ToArray()).ToArray();
+            var data = raw.Select(l => l.Split(',').Select(double.Parse).ToArray()).ToArray();
             var dataX = data.SelectMany(l => l.Take(64)).ToArray();
             var dataY = Categorical(data.Select(l => (int)l[64]).ToArray());
 
@@ -34,10 +34,10 @@ namespace DesertLand
             int splitX = (int)(lX * ratio);
             int splitY = (int)(lY * ratio);
 
-            float[] dataTrainX = dataX.Take(splitX * 64).Select(i => i / 16.0f).ToArray();
-            float[] dataTestX = dataX.Skip(splitX * 64).Select(i => i / 16.0f).ToArray();
-            float[] dataTrainY = dataY.Take(splitY * 10).ToArray();
-            float[] dataTestY = dataY.Skip(splitY * 10).ToArray();
+            double[] dataTrainX = dataX.Take(splitX * 64).Select(i => i / 16.0).ToArray();
+            double[] dataTestX = dataX.Skip(splitX * 64).Select(i => i / 16.0).ToArray();
+            double[] dataTrainY = dataY.Take(splitY * 10).ToArray();
+            double[] dataTestY = dataY.Skip(splitY * 10).ToArray();
 
             var trainX = ND.CreateNDarray(dataTrainX, splitX, 64).CastCopy<Type>();
             var trainY = ND.CreateNDarray(dataTrainY, splitY, 10).CastCopy<Type>();
